+++ app/resources/views/layouts/topnav.blade.php
@@ -7,6 +7,7 @@
 
     <title>@yield('title', 'OpenDominion')</title>
 
+    <link rel="author" href="{{ asset('humans.txt') }}" />
     <meta name="csrf-token" content="{{ csrf_token() }}">
 
     <meta property="og:title" content="OpenDominion">